Modularize the back-end codebase, movies, weather and handleError modules
