+++ modules/errorHandler.js
@@ -1,0 +1,5 @@
+function errorHandler(error, res) {
+  res.status(500).send({ error: "Somthing Wrong!!" });
+}
+
+module.exports = errorHandler;
